hack to make PskKeyManager work on newer Android
